chore: remove extension hook & rename chaindata provider methods
