+++ packages/balances-react/src/hooks/useBalancesHydrate.ts
@@ -3,7 +3,6 @@ import { useMemo } from "react"
 
 import { useChains } from "./useChains"
 import { useEvmNetworks } from "./useEvmNetworks"
-import useExtensionChaindataSyncEffect from "./useExtensionChaindataSyncEffect"
 import { useTokenRates } from "./useTokenRates"
 import { useTokens } from "./useTokens"
 import { useWithTestnets } from "./useWithTestnets"
@@ -14,8 +13,6 @@ export const useBalancesHydrate = (): HydrateDb => {
   const evmNetworks = useEvmNetworks(withTestnets)
   const tokens = useTokens(withTestnets)
   const tokenRates = useTokenRates()
-
-  useExtensionChaindataSyncEffect()
 
   return useMemo(
     () => ({ chains, evmNetworks, tokens, tokenRates }),
